% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddFun.R
\name{dgammaIDD}
\alias{dgammaIDD}
\title{IDD transmissibility curve using the gamma PDF}
\usage{
dgammaIDD(x, params, getParams = F)
}
\arguments{
\item{x}{A vector ranging from 1 to the maximum length of the infectious period.}

\item{params}{A list giving parameter values for the shape and rate of the gamma PDF.}

\item{getParams}{logical. Tells the function to return the parameter names.}
}
\value{
Gamma PDF
\code{dgamma(x, shape = params$shape, rate = params$rate)}
}
\description{
IDD transmissibility curve using the gamma PDF
}
\examples{
dgammaIDD(1:15, params = list(shape = 4, rate = 1))
}
