% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddFun.R
\name{dlnormIDD}
\alias{dlnormIDD}
\title{IDD transmissibility curve using the log normal PDF}
\usage{
dlnormIDD(x, params, getParams = F)
}
\arguments{
\item{x}{A vector ranging from 1 to the maximum length of the infectious period.}

\item{params}{A list giving parameter values for the meanlog and sdlog of the log normal PDF.}

\item{getParams}{logical. Tells the function to return the parameter names.}
}
\value{
Log normal PDF
\code{dlnorm(x, meanlog = params$meanlog, sdlog = params$sdlog)}
}
\description{
IDD transmissibility curve using the log normal PDF
}
\examples{
dlnormIDD(1:15, params = list(meanlog = 2, sdlog = 1))
}
