% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddFun.R
\name{logitIDD}
\alias{logitIDD}
\title{IDD transmissibility curve using logistic decay}
\usage{
logitIDD(x, params, getParams = F)
}
\arguments{
\item{x}{A vector ranging from 1 to the maximum length of the infectious period.}

\item{params}{A list giving parameter values for the decay rate \code{rate}
and inflection point \code{mid} of the logistic decay function.}

\item{getParams}{logical. Tells the function to return the parameter names.}
}
\value{
\code{1 / (1 + exp(params$rate * (x - params$mid)))}
}
\description{
IDD transmissibility curve using logistic decay
}
\examples{
logitIDD(1:15, params = list(mid = 8, rate = 1.5))
}
