% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{psProbVec}
\alias{psProbVec}
\title{Get path-specific vector of removal probabilities over duration of infection}
\usage{
psProbVec(maxInf, dist, psParams)
}
\arguments{
\item{maxInf}{Integer. The maximum length of time an individual can
spend in the infectious compartment}

\item{dist}{Character. Specifies the distribution used to describe the duration
of the infectious period.  One of 'exp', 'gamma', or 'weibull'}

\item{psParams}{List. the parameters associated with the distribution specified.}
}
\value{
The path-specific vector of removal probabilities.
}
\description{
Get path-specific vector of removal probabilities over duration of infection
}
\examples{
psProbVec(1:14, 'gamma', list(shape = 4, rate = 1))
}
