% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iddFun.R
\name{splineIDD}
\alias{splineIDD}
\title{IDD transmissibility curve using a spline basis}
\usage{
splineIDD(x, params, XBasis, getParams = F)
}
\arguments{
\item{x}{A vector ranging from 1 to the maximum length of the infectious period.}

\item{params}{A list giving coefficients for each basis in the basis matrix,
named as \code{b1}, \code{b2}, ...}

\item{XBasis}{Basis matrix. Generally the output from something such as \code{splines::bs()}.}

\item{getParams}{logical. Tells the function to return the parameter names.}
}
\value{
\code{exp(XBasis \%*\% b)}, where \code{b = c(params$b1, params$b2, ...)}
}
\description{
IDD transmissibility curve using a spline basis
}
\details{
The spline IDD curve is exponentiated to ensure f(x) > 0
}
\examples{
splineIDD(1:15, params = list(b1 = 3, b2 = 1, b3 = -3),
                XBasis = splines::bs(1:15, df = 3))
}
