% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{psProbVec}
\alias{psProbVec}
\title{Get path-specific vector of removal probabilities over duration of infection}
\usage{
psProbVec(maxInf, dist, psParams)
}
\arguments{
\item{maxInf}{Integer. The maximum length of time an individual can
spend in the infectious compartment}

\item{dist}{Character.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Get path-specific vector of removal probabilities over duration of infection
}
\examples{
add(1, 1)
add(10, 1)
}
